chatbot_ai/requirements.txt
fastapi
uvicorn
langchain
openai
python-dotenv
faiss-cpu
pypdf
tiktoken
chromadb